Rename files and links
