+++ src/layouts/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
 import { ThemeProvider } from 'styled-components'
 import { Provider } from '../components/AppContext'
-import Navigation from '../components/navigation/Navigation'
+import Navigation from '../components/navigation'
 import theme from '../theme/theme'
 
 import './index.css'
